# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FTC Events API",
...
    "version": "v2.0"
  },
  "servers": [
    {
      "url": "http://ftc-api.firstinspires.org"
    }
  ],
  "paths": {
    "/v2.0/{season}/advancement/{eventCode}": {
      "get": {
        "tags": [
          "Advancement"
        ],
        "summary": "Event Advancement",
        "description": "The event advancement endpoint returns details about teams advancing from a particular event in a particular season.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year of the event from which the event advancement is requested. Must be 4 digits > 2022",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventCode",
            "in": "path",
            "description": "Case insensitive alphanumeric `eventCode` of the event from which the advancement results are requested. Must be at least 3 characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excludeSkipped",
            "in": "query",
            "description": "`excludeSkipped=true` to exclude skipped advancement slots. Slots are skipped if no team meets the criteria, the team has already advanced, the team declined, or the team was ineligible.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancementModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
    "/v2.0/{season}/advancement/{eventCode}/source": {
      "get": {
        "tags": [
          "Advancement"
        ],
        "summary": "Advancement Source",
        "description": "The advancement source API returns details about where teams advanced to a specified event from.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year of the event from which the advancement is requested. Must be 4 digits >= 2022",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventCode",
            "in": "path",
            "description": "Case insensitive alphanumeric `eventCode` of the event for which teams advanced to. Must be at least 3 characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDeclines",
            "in": "query",
            "description": "`includeDeclines=true` to include teams that declined their advancement slot to this event.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdvancementSourceModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
    "/v2.0/{season}/leagues": {
      "get": {
        "tags": [
          "Leagues"
        ],
        "summary": "League Listings",
        "description": "The league listings API returns all FTC leagues in a particular season. You can specify a `regionCode` to filter to leagues within a particular region. To filter to a specific league, supply both a `regionCode` and a `leagueCode`. The returned objects have a `parentLeagueCode` field, which indicates the league is a child league if not null and provides the code of the parent league. The `regionCode` of the parent league will always match the child.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year from which the league listings are requested. Must be 4 digits",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "regionCode",
            "in": "query",
            "description": "Case-sensitive alphanumeric `regionCode` of a region to filter for.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "leagueCode",
            "in": "query",
            "description": "Case-sensitive alphanumeric `leagueCode` of the league within the specified region to query.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonLeagueListingsModel_Version2"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
    "/v2.0/{season}/leagues/members/{regionCode}/{leagueCode}": {
      "get": {
        "tags": [
          "Leagues"
        ],
        "summary": "League Membership",
        "description": "The league membership API returns the list of team numbers for the teams that are members of a particular league. Leagues are specified by a `regionCode` in combination with a `leagueCode`.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year. Must be 4 digits",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "regionCode",
            "in": "path",
            "description": "Case sensitive alphanumeric `regionCode` of the region the league belongs to.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "leagueCode",
            "in": "path",
            "description": "Case sensitive alphanumeric `leagueCode` of the league.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueMemberListModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
    "/v2.0/{season}/leagues/rankings/{regionCode}/{leagueCode}": {
      "get": {
        "tags": [
          "Leagues"
        ],
        "summary": "League Rankings",
        "description": "The league rankings API returns team ranking detail from a particular league in a particular season. League rankings are only the cumulative rankings from League Meets - they do not include performance at the League Tournament. To get League Tournament Rankings, use the Event Rankings endpoint.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year. Must be 4 digits",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "regionCode",
            "in": "path",
            "description": "Case sensitive alphanumeric `regionCode` of the region the league belongs to.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "leagueCode",
            "in": "path",
            "description": "Case sensitive alphanumeric `leagueCode` of the league.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRankingsModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
...
    "/v2.0/{season}/events": {
      "get": {
        "tags": [
          "Season Data"
        ],
        "summary": "Event Listings",
        "description": "The event listings API returns all FTC official regional events in a particular season. You can specify an `eventCode` if you would only like data about one specific event. If you specify an `eventCode` you cannot specify any other optional parameters. Alternately, you can specify a `teamNumber` to retrieve only the listings of events being attended by the particular team. If you specify a `teamNumber` you cannot specify an `eventCode`.\n\nThe response for event listings contains a special field called divisionCode. For example, the FIRST Championship contains two Divisions. As an example of a response, the event listings for a Division will have a divisionCode that matches the FIRST Championship event code (as they are divisions of that event). This allows you to see the full structure of events, and how they relate to each other.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year from which the event listings are requested. Must be 4 digits",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventCode",
            "in": "query",
            "description": "Case insensitive alphanumeric `eventCode` of the event about which details are requested.",
            "schema": {
              "type": "string",
              "default": "0"
            }
          },
          {
            "name": "teamNumber",
            "in": "query",
            "description": "Numeric `teamNumber` of the team from which the attending event listings are requested.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonEventListingsModel_Version2"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
    "/v2.0/{season}/teams": {
      "get": {
        "tags": [
          "Season Data"
        ],
        "summary": "Team Listings",
        "description": "The team listings API returns all FTC official teams in a particular season. If specified, the `teamNumber` parameter will return only one result with the details of the requested `teamNumber`. Alternately, the `eventCode` parameter allows sorting of the team list to only those teams attending a particular event in the particular season. If you specify a teamNumber parameter, you cannot additionally specify an `eventCode` and/or `state` in the same request, or you will receive an HTTP 501. If you specify the `state` parameter, it should be the full legal name of the US state or international state/prov, such as New Hampshire or Ontario. Values on this endpoint are \"pass through\" values from the TIMS registration system. As such, if the team does not specify a value for a field, it may be presented in the API as null.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year from which the team listings are requested. Must be 4 digits.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "teamNumber",
            "in": "query",
            "description": "Numeric `teamNumber` of the team about which information is requested. Must be 1 to 5 digits.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "eventCode",
            "in": "query",
            "description": "Case insensitive alphanumeric `eventCode` of the event from which details are requested.",
            "schema": {
              "type": "string",
              "default": "0"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "Full legal name of the US state or international state/prov",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Numeric page of results to return.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonTeamListingsModel_Version2"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
    "/v2.0/{season}/schedule/{eventCode}/{tournamentLevel}/hybrid": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "summary": "Hybrid Schedule",
        "description": "The schedule API returns the match schedule for the desired tournament level of a particular event in a particular season in the hybrid format. When a match has been played, the match result related details will be filled. When a match has not yet happened, match result related fields will be null. All parameters, except start and end, are required for the hybrid schedule.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year of the event from which the hybrid schedule is requested. Must be 4 digits",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventCode",
            "in": "path",
            "description": "Case insensitive alphanumeric eventCode of the event from which the hybrid schedule is requested. Must be at least 3 characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tournamentLevel",
            "in": "path",
            "description": "Required tournamentLevel of desired score details.",
            "required": true,
            "schema": {
              "enum": [
                "qual",
                "playoff"
              ],
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "`start` match number for subset of results to return (inclusive).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "`end` match number for subset of results to return (inclusive).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 999
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventScheduleHybridModel_Version2"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
    "/v2.0/{season}/schedule/{eventCode}": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "summary": "Event Schedule",
        "description": "The schedule API returns the match schedule for the desired tournament level of a particular event in a particular season. You must also specify a `tournamentLevel` from which to return the results. Alternately, you can specify a `teamNumber` to filter the results to only those in which a particular team is participating. There is no validation that the `teamNumber` you request is actually competing at the event, if they are not, the response will be empty. You can also specify the parameters together, but cannot make a request without at least one of the two.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year of the event from which the schedule is requested. Must be 4 digits",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventCode",
            "in": "path",
            "description": "Case insensitive alphanumeric `eventCode` of the event from which the schedule are requested. Must be at least 3 characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tournamentLevel",
            "in": "query",
            "description": "Required tournamentLevel of desired score details.",
            "schema": {
              "enum": [
                "qual",
                "playoff"
              ],
              "type": "string",
              "default": "0"
            }
          },
          {
            "name": "teamNumber",
            "in": "query",
            "description": "`teamNumber` to search for within the schedule. Only returns matches in which the requested team participated.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "`start` match number for subset of results to return (inclusive).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "`end` match number for subset of results to return (inclusive).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 999
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventScheduleModel_Version2"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
    "/v2.0/{season}/rankings/{eventCode}": {
      "get": {
        "tags": [
          "Rankings"
        ],
        "summary": "Event Rankings",
        "description": "The rankings API returns team ranking detail from a particular event in a particular season. Optionally, the `top` parameter can be added to the query string to request a subset of the rankings based on the highest ranked teams at the time of the request. Alternately, you can specify the `teamNumber` parameter to retrieve the ranking on one specific team. You cannot specify both a `top` and `teamNumber` in the same call.",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "description": "Numeric year of the event from which the rankings are requested. Must be 4 digits",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventCode",
            "in": "path",
            "description": "Case insensitive alphanumeric `eventCode` of the event from which the rankings are requested. Must be at least 3 characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teamNumber",
            "in": "query",
            "description": "Team number of the team whose ranking is requested.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "top",
            "in": "query",
            "description": "number of requested `top` ranked teams to return in result.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRankingsModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvancementModel": {
        "title": "Advancement",
        "type": "object",
        "properties": {
          "advancesTo": {
            "type": "string",
            "description": "The event code of the event this event advanced to.",
            "nullable": true
          },
          "slots": {
            "type": "integer",
            "description": "The number of teams that advanced.",
            "format": "int32"
          },
          "advancement": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdvancementSlot"
            },
            "description": "The ordered list of advancement criteria and teams that met them. Null if the event is unpublished.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AdvancementSlot": {
        "title": "Advancement Order",
        "type": "object",
        "properties": {
          "team": {
            "type": "integer",
            "description": "The team number of the team earning this advancement slot. Null if no team meets the criteria.",
            "format": "int32",
            "nullable": true
          },
          "teamId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "teamProfileId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "teamInternalId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "displayTeam": {
            "type": "string",
            "description": "The display-friendly string representation of the number of the team earning this advancement slot. Null if no team meets the criteria.",
            "nullable": true
          },
          "slot": {
            "type": "integer",
            "description": "The number of the advancement criteria",
            "format": "int32"
          },
          "criteria": {
            "type": "string",
            "description": "String describing the advancement criteria",
            "nullable": true
          },
          "declined": {
            "type": "boolean",
            "description": "true if the team declined their advancement invitation"
          },
          "status": {
            "$ref": "#/components/schemas/ApiAdvancementStatus"
          }
        },
        "additionalProperties": false
      },
      "AdvancementSourceModel": {
        "title": "Advanced",
        "type": "object",
        "properties": {
          "advancedFrom": {
            "type": "string",
            "description": "The event code of the event this team advanced from. Null if unknown.",
            "nullable": true
          },
          "advancedFromRegion": {
            "type": "string",
            "description": "The region code of the region this team advanced from. Null if target event is a region-level event.",
            "nullable": true
          },
          "slots": {
            "type": "integer",
            "description": "The number of teams that advanced from the source event.",
            "format": "int32"
          },
          "advancement": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdvancementSlot"
            },
            "description": "The ordered list of advancement criteria and teams that met them. Null if the event is unpublished.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EventRankingsModel": {
        "title": "Event Ranking List",
        "type": "object",
        "properties": {
          "rankings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamRankingModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventScheduleHybridModel_Version2": {
        "title": "Hybrid Schedule",
        "type": "object",
        "properties": {
          "schedule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleHybridModel_Version2"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventScheduleModel_Version2": {
        "title": "Scheduled Match List",
        "type": "object",
        "properties": {
          "schedule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduledMatchModel_Version2"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LeagueMemberListModel": {
        "title": "League Members",
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SeasonEventListingsModel_Version2": {
        "title": "Event List",
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeasonEventModel_Version2"
            },
            "nullable": true
          },
          "eventCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SeasonLeagueListingsModel_Version2": {
        "title": "League List",
        "type": "object",
        "properties": {
          "leagues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeasonLeagueModel_Version2"
            },
            "nullable": true
          },
          "leagueCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SeasonLeagueModel_Version2": {
        "title": "League",
        "type": "object",
        "properties": {
          "region": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "remote": {
            "type": "boolean",
            "nullable": true
          },
          "parentLeagueCode": {
            "type": "string",
            "nullable": true
          },
          "parentLeagueName": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SeasonTeamListingsModel_Version2": {
        "title": "Team List",
        "type": "object",
        "properties": {
          "teams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeasonTeamModel_Version2"
            },
            "nullable": true
          },
          "teamCountTotal": {
            "type": "integer",
            "format": "int32"
          },
          "teamCountPage": {
            "type": "integer",
            "format": "int32"
          },
          "pageCurrent": {
            "type": "integer",
            "format": "int32"
          },
          "pageTotal": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "basic": {
        "type": "http",
        "description": "To register for the api, register at https://ftc-events.firstinspires.org/services/API/register",
        "scheme": "basic"
      }
    }
  }
}